backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0,<5.0.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0
sqlalchemy>=2.0.0
pandas>=2.0.0
openpyxl>=3.1.0
pymongo>=4.6.0
psycopg2-binary>=2.9.0
pymysql>=1.1.0
python-dotenv>=1.0.0
google-cloud-firestore>=2.11.0
google-auth>=2.23.0

